update the databse column type

+++ backend/src/migrations/003_create_subsidiaries_table.ts
@@ -3,8 +3,8 @@ import { Kysely, sql } from 'kysely';
 export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
     .createTable('subsidiaries')
-    .addColumn('id', 'serial', (col) => col.primaryKey())
-    .addColumn('corporate_id', 'integer', (col) => col.notNull().references('corporates.id').onDelete('cascade'))
+    .addColumn('id', 'bigserial', (col) => col.primaryKey())
+    .addColumn('corporate_id', 'bigint', (col) => col.notNull().references('corporates.id').onDelete('cascade'))
     .addColumn('company_name', 'varchar(255)', (col) => col.notNull())
     .addColumn('reg_number', 'varchar(50)', (col) => col.notNull())
     .addColumn('office_address1', 'varchar(255)', (col) => col.notNull())
